+++ app/shared/com/ortb/model/reflection/ConstructorWrapperModel.scala
@@ -1,4 +1,4 @@
-package shared.com.ortb.model
+package shared.com.ortb.model.reflection
 
 import java.lang.reflect.Constructor
