+++ app/shared/com/ortb/model/reflection/FieldWrapperModel.scala
@@ -1,4 +1,4 @@
-package shared.com.ortb.model
+package shared.com.ortb.model.reflection
 
 import java.lang.reflect.Field
