+++ app/shared/com/ortb/model/reflection/ModifierEnhancementConcreteImplementation.scala
@@ -1,4 +1,6 @@
-package shared.com.ortb.model
+package shared.com.ortb.model.reflection
+
+import shared.com.ortb.model.traits.ModifierEnhancement
 
 class ModifierEnhancementConcreteImplementation
 (val modifier : Int)
