+++ app/shared/com/ortb/model/results/ApiPaginationResultsModel.scala
@@ -1,6 +1,6 @@
 package shared.com.ortb.model.results
 
-import shared.com.ortb.model.attributes.{ ApiPaginationResponseAttributesModel, GenericControllerResponseAttributesModel }
+import shared.com.ortb.model.attributes.ApiPaginationResponseAttributesModel
 
 case class ApiPaginationResultsModel[TRow, TKey](
   attributes : Option[ApiPaginationResponseAttributesModel] = None,
