+++ app/shared/com/ortb/model/results/ResultModel.scala
@@ -1,0 +1,5 @@
+package shared.com.ortb.model.results
+
+abstract class ResultModel[T] (
+  result : Option[T]
+)
