+++ app/shared/com/ortb/model/results/ResultWithSuccessModel.scala
@@ -2,4 +2,4 @@ package shared.com.ortb.model.results
 
 case class ResultWithSuccessModel[T](
   result : Option[T],
-  isSuccess : Boolean)
+  isSuccess : Boolean) extends ResultModel[T](result)
