+++ app/shared/com/ortb/model/traits/ItemsExistence.scala
@@ -1,4 +1,4 @@
-package shared.com.ortb.model
+package shared.com.ortb.model.traits
 
 trait ItemsExistence[T] {
   val hasItem : Boolean
